refactor: rename apps/chat -> apps/canopy_sessions (canonical session system)
Pure rename, no behavior change. Django's app label had to change too
(chat -> canopy_sessions) since plain "sessions" collides with
django.contrib.sessions's built-in label — that's why the app was
originally named chat (see ARCHITECTURE.md). Supersedes the task-6
brief's "apps/sessions" target with the corrected name.

- git mv apps/chat -> apps/canopy_sessions; apps.py label + config
  class renamed (ChatConfig -> CanopySessionsConfig).
- INSTALLED_APPS, asgi.py, api.py router import, harness FK string ref
  ("chat.Session" -> "canopy_sessions.Session", incl. the migration
  0013 to='chat.session' FK target), and every migration dependency
  tuple updated to the new label.
- All apps.chat imports across tests + services updated.
- Deliberately left untouched: the ws/chat/ Channels protocol strings
  (chat.send, chat.stream_*, chat.turn_event, group name "chat.{id}")
  and the /api/chat URL prefix — those are ace-web's canonical wire
  protocol, not the app label, and their rename is a separate later
  plan.
- Migration mechanism: since a fresh test DB replays migrations under
  the new label directly (no chat_* tables ever exist), no RunSQL
  table-rename migration was needed — the graph is self-consistent
  under canopy_sessions and makemigrations --check reports no changes.
