+++ ARCHITECTURE.md
@@ -41,7 +41,7 @@ the arrow's direction and enforce it in CI.
 | `system` | **framework** | System metadata / AI-backend status. |
 | `push` | **framework** | Web Push subscription registry (VAPID keypair + `PushSubscription` rows). Agent-agnostic — any agent's board could trigger a send; observes `agents` (never the reverse), same direction `harness` takes. |
 | `realtime` | **framework** | WebSocket transport (Django Channels + Redis): live-tails the `harness` `TurnEvent` ledger (`turn.{id}`) and pushes `/supervisor` runner-status + waiting-count deltas. Fan-out mirrors `push` (signal/`post_save` → `on_commit` → `group_send`); observes `harness`/`push`/`agents`, never the reverse. Wave 4 (`docs/superpowers/specs/2026-07-16-realtime-chat-cloud-runner-program-design.md`). |
-| `chat` | **framework** | Live, multiplayer chat sessions — the interactive front-door to a durable `harness` Turn. A `Session` "send" enqueues a session-target Turn; the assistant stream lands in the `TurnEvent` ledger and is projected into `Message` rows. SP3 adds co-edited `Draft` + `SessionParticipant` + cache-backed presence and a per-session `SessionConsumer` (`ws/chat/{id}/`) over the `realtime` transport. Agent-agnostic (opaque `metadata` for product linkage). Named `chat` (not `sessions`) to avoid the `django.contrib.sessions` label + `session_sharing` route collisions. Wave 4 SP2–SP3. |
+| `canopy_sessions` | **framework** | Live, multiplayer chat sessions — the interactive front-door to a durable `harness` Turn. A `Session` "send" enqueues a session-target Turn; the assistant stream lands in the `TurnEvent` ledger and is projected into `Message` rows. SP3 adds co-edited `Draft` + `SessionParticipant` + cache-backed presence and a per-session `SessionConsumer` (`ws/chat/{id}/`) over the `realtime` transport. Agent-agnostic (opaque `metadata` for product linkage). Named `canopy_sessions` — plain `sessions` collides with the `django.contrib.sessions` label (and `session_sharing` already owns the shared-transcript name). The `/api/chat` route prefix and `ws/chat/` protocol strings are unchanged (a later plan renames those). Wave 4 SP2–SP3. |
 | `projects` | **product** | Canopy's portfolio/insights feature: repos + which canopy skills ran (`skills[]`, `skill_name`, hygiene-skill frontend). Not a generic registry today — promote to framework only when a real second consumer needs one. |
 | `walkthroughs` | **product** | DDD walkthrough artifacts (HTML/video demos). |
 | `reviews` | **product** | DDD narrative review surface. |
