+++ CLAUDE.md
@@ -14,7 +14,7 @@ first-class AI agents, demo-driven development (DDD), walkthroughs, and shareout
 - **AI:** Anthropic Claude API via SSE streaming. Dual backend — direct API key (`AI_BACKEND=api`) or Claude Code CLI subscription (`AI_BACKEND=cli`), switchable at runtime via `/api/ai/switch/`.
 - **MCP server:** `apps/mcp/` is a FastMCP 3.x Streamable-HTTP server mounted into the ASGI app at `/api/mcp/` (wired in `config/asgi.py`). Tools run **as the authenticated user** via per-user PAT (`CanopyPATVerifier`) and reuse the same service functions as the REST views, so the two surfaces can't drift. See `docs/architecture/mcp-surface.md`.
 - **Deployment:** AWS ECS Fargate on the shared labs platform (account `858923557655`, `us-east-1`), served at `https://labs.connect.dimagi.com/canopy/` behind the shared ALB. One container (Django serves the built SPA + API + MCP). Deploys run from GitHub via the **Deploy to Labs (AWS)** workflow (`.github/workflows/deploy-labs.yml`): build → push to ECR (`labs-jj-canopy-web`) → register task-def revision → auto-migrate (idempotent; skippable only via `skip_migrations`) → roll the ECS service. Infra is provisioned by `deploy/aws/canopy-web.cfn.yaml` (CloudFormation). Runtime settings in `config/settings/connectlabs.py` (extends `production.py`; `FORCE_SCRIPT_NAME=/canopy`, shared RDS `canopy_web` DB).
-- **Framework/product boundary (the one invariant):** apps split into **framework** (generic, agent-agnostic substrate — `agents`, `agent_runs`, `workspaces`, `api`, `common`, `timeline`, `tokens`, `session_sharing`, `issues`, `mcp`, `system`, `realtime` (Channels WS transport), `chat` (multiplayer chat sessions)) and **product** (canopy's own features — `projects`, `walkthroughs`, `reviews`, `shareouts`, `runs`). **Framework code must never import product code; product freely imports framework.** This keeps the blend cuttable (the framework apps could lift onto a standalone host without dragging canopy's product). It's a *direction, not a wall* — we don't move apps into `framework/`/`product/` folders. Enforced by `tests/test_architecture_boundary.py` (fails CI on a framework→product import, or on a new app left untiered). Full rationale, the per-app tier table, and the accepted carve-outs (the `api` composition root, the `mcp` insights tool): **`ARCHITECTURE.md`**. The framework apps are being harvested as the generic layer out of ACE — see `docs/superpowers/specs/2026-06-24-canopy-framework-harvest-design.md`.
+- **Framework/product boundary (the one invariant):** apps split into **framework** (generic, agent-agnostic substrate — `agents`, `agent_runs`, `workspaces`, `api`, `common`, `timeline`, `tokens`, `session_sharing`, `issues`, `mcp`, `system`, `realtime` (Channels WS transport), `canopy_sessions` (multiplayer chat sessions)) and **product** (canopy's own features — `projects`, `walkthroughs`, `reviews`, `shareouts`, `runs`). **Framework code must never import product code; product freely imports framework.** This keeps the blend cuttable (the framework apps could lift onto a standalone host without dragging canopy's product). It's a *direction, not a wall* — we don't move apps into `framework/`/`product/` folders. Enforced by `tests/test_architecture_boundary.py` (fails CI on a framework→product import, or on a new app left untiered). Full rationale, the per-app tier table, and the accepted carve-outs (the `api` composition root, the `mcp` insights tool): **`ARCHITECTURE.md`**. The framework apps are being harvested as the generic layer out of ACE — see `docs/superpowers/specs/2026-06-24-canopy-framework-harvest-design.md`.
 
 ## Development
 
@@ -284,7 +284,7 @@ The Item **carries its own text** (message semantics, like an email) rather than
 
 **The inbox is a pure `Item` query — no projections.** The supervisor inbox (`/supervisor`) and the per-agent **Inbox** rail both render `Item.filter(state=open)`, decidable in place (`decide`/`dismiss` inline). The old `needs_you` aggregation and its projections (`SUGGESTED`/human-blocked tasks, run gates/failed steps, the schedule nag) were **deleted** — "needs you" was never a first-class concept, just a label on a function. Producers now raise real Items: the **schedule nag** is server-local (an unattended grace-released occurrence raises a `review` Item whose `implement` re-runs the schedule — `services._raise_schedule_nag`, dismissed on a later `DONE`); **run gates** and **task decisions** are raised by their producers (the runner's `reviews.py`, the fleet `task-tracker` skill) as follow-on repo work — the inbox simply shows whatever Items exist. The task board (`AgentTask`) stays as the "who has the ball" surface but no longer feeds the inbox. See `docs/superpowers/specs/2026-07-21-supervisor-inbox-items-only-design.md`.
 
-### Chat (`apps/chat`) — multiplayer chat sessions (the live front-door)
+### Chat (`apps/canopy_sessions`) — multiplayer chat sessions (the live front-door)
 A `Session` is a durable conversation **with an agent** (agent-agnostic, workspace-tenanted). A send commits the co-edited `Draft` → enqueues a **session `Turn`** (a third `Turn` target: agent XOR project XOR session) → a session-capable runner executes it and the reply streams back over the ledger; `Message` rows are a materialized projection. The per-session WebSocket `SessionConsumer` (`ws/chat/{id}/`) carries the connect snapshot, co-edited draft (version guard + derived soft-lock), presence, and the streamed reply — speaking ace-web's **canonical protocol** (`session.state`/`chat.stream_*`/`draft.*`/`presence.*`) so the shared `canopy-ui/chat` kit drives it and ace-web can adopt the same kit. Fan-out is generic (`apps/realtime`); the consumer translates `TurnEvent`s → `chat.stream_*`.
 - `POST /api/chat/` — Create a session (`agent_slug`, `title`, `metadata`); tenant route `/api/w/{ws}/chat/` creates in that workspace (cross-workspace new-chat).
 - `GET /api/chat/` — List my sessions (creator-scoped; flat mount = across all my workspaces).
