+++ apps/canopy_sessions/apps.py
@@ -1,10 +1,10 @@
 from django.apps import AppConfig
 
 
-class ChatConfig(AppConfig):
+class CanopySessionsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "apps.chat"
-    label = "chat"
+    name = "apps.canopy_sessions"
+    label = "canopy_sessions"
 
     def ready(self) -> None:
         from . import signals  # noqa: F401  (Message projection receiver)
