+++ apps/canopy_sessions/consumers.py
