+++ apps/canopy_sessions/participants.py
