+++ apps/canopy_sessions/presence.py
