+++ tests/test_chat_executor.py
@@ -5,8 +5,8 @@ import pytest
 from django.contrib.auth.models import User
 
 from apps.agents.models import Agent
-from apps.chat import services as chat
-from apps.chat.executor import execute_turn_stub
+from apps.canopy_sessions import services as chat
+from apps.canopy_sessions.executor import execute_turn_stub
 from apps.harness.models import Turn
 from apps.workspaces.models import Workspace
