+++ tests/test_chat_send.py
@@ -5,8 +5,8 @@ import pytest
 from django.contrib.auth.models import User
 
 from apps.agents.models import Agent
-from apps.chat import services as chat
-from apps.chat.models import Session
+from apps.canopy_sessions import services as chat
+from apps.canopy_sessions.models import Session
 from apps.harness import services as harness
 from apps.harness.models import Turn
 from apps.workspaces.models import Workspace
