+++ tests/test_harness_claim_sessions.py
@@ -11,7 +11,7 @@ import pytest
 from django.contrib.auth import get_user_model
 from django.utils import timezone
 
-from apps.chat.models import Session
+from apps.canopy_sessions.models import Session
 from apps.harness import services
 from apps.harness.models import Runner, Turn
 from apps.workspaces.models import Workspace, WorkspaceMembership
